refactor(onchain-restructure): 네트워크 관련 로직 디렉터리 구조 변경
