+++ lib/services/model/response/electrum_response_types.dart
