+++ lib/services/model/stream/base_stream_state.dart
