+++ lib/services/network/node_client.dart
@@ -1,10 +1,10 @@
 import 'package:coconut_lib/coconut_lib.dart' as lib;
 import 'package:coconut_wallet/model/utxo/utxo.dart';
 import 'package:coconut_wallet/model/wallet/balance.dart';
-import 'package:coconut_wallet/services/network/dto/block_timestamp.dart';
-import 'package:coconut_wallet/services/network/dto/fetch_transaction_response.dart';
-import 'package:coconut_wallet/services/network/electrum/electrum_api.dart';
-import 'package:coconut_wallet/services/network/stream/base_stream_state.dart';
+import 'package:coconut_wallet/services/model/response/block_timestamp.dart';
+import 'package:coconut_wallet/services/model/response/fetch_transaction_response.dart';
+import 'package:coconut_wallet/services/electrum_service.dart';
+import 'package:coconut_wallet/services/model/stream/base_stream_state.dart';
 
 /// @nodoc
 abstract class NodeClient {
@@ -52,6 +52,6 @@ abstract class NodeClientFactory {
 class ElectrumNodeClientFactory implements NodeClientFactory {
   @override
   Future<NodeClient> create(String host, int port, {bool ssl = true}) async {
-    return ElectrumApi.connectSync(host, port, ssl: ssl);
+    return ElectrumService.connectSync(host, port, ssl: ssl);
   }
 }
